content-security-policy/navigation/to-javascript-parent-initiated-parent-csp.html

<!DOCTYPE html>
<head>
<meta http-equiv="content-security-policy" content="script-src 'self' 'nonce-abc'">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<!-- <frame-with-csp.sub.html> without parameters corresponds to no csp.-->
<iframe src="support/frame-with-csp.sub.html"></iframe>
<script nonce='abc'>
  function addSuccessAndFailureEventListeners(test, resolve) {
    window.addEventListener("message", test.step_func(function(e) {
      if (e.data == "executed")
        assert_true(false, "Javascript url executed");
    }));
    window.addEventListener('securitypolicyviolation', test.step_func_done(function(e) {
      assert_equals(e.blockedURI, 'inline');
      assert_equals(e.violatedDirective, 'script-src-elem');
      resolve();
    }));
  }

  promise_test(t => new Promise(resolve => {
    addSuccessAndFailureEventListeners(t, resolve);
    frames[0].location.href = "javascript:parent.postMessage('executed', '*')";
  }), "Should not have executed the javascript url");
</script>
</body>
