LayoutTests/imported/w3c/web-platform-tests/content-security-policy/navigation/to-javascript-parent-initiated-parent-csp-expected.txt



FAIL Should not have executed the javascript url assert_equals: expected "inline" but got ""
